+++ src/Streamly/Time.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE FlexibleContexts #-}
+
 -- |
 -- Module      : Streamly.Time
 -- Copyright   : (c) 2017 Harendra Kumar
@@ -10,16 +12,26 @@
 -- Time utilities for reactive programming.
 
 module Streamly.Time
-{-# DEPRECATED
-   "Please use the \"rate\" combinator instead of the functions in this module"
-  #-}
-    ( periodic
+    ( intervalsOf
+
+    -- * Deprecated
+    , periodic
     , withClock
     )
 where
 
-import Control.Monad (when)
 import Control.Concurrent (threadDelay)
+import Control.Monad (when)
+import Control.Monad.IO.Class (MonadIO(liftIO))
+import Data.Maybe (fromJust, isJust)
+
+import Streamly (parallel)
+import Streamly.SVar (MonadAsync)
+import Streamly.Streams.StreamK (IsStream(..))
+import Streamly.Fold (Fold)
+
+import qualified Streamly.Fold as FL
+import qualified Streamly.Prelude as S
 
 -- | Run an action forever periodically at the given frequency specified in per
 -- second (Hz).
@@ -72,3 +84,42 @@ withClock clock freq action = do
             lastPeriod = (baseTime - lastAdj) `div` freq
             newDelay   = max 0 (delay + period - lastPeriod)
         return (baseTime, newTick, newDelay)
+
+-------------------------------------------------------------------------------
+-- Stream element unaware: Time based buffering
+-------------------------------------------------------------------------------
+
+-- There are two ways in which time based buffering can be implemented.
+--
+-- 1) One is to use an API like groupsOf to group elements by time and then
+-- flush when the group is complete. This will require checking time on each
+-- element yield.
+--
+-- 2) Use an async thread to write. The producer thread would just queue the
+-- elements to be written on a queue associated with the writer thread. When
+-- the write thread is scheduled elements that have been buffered will be
+-- written based on a buffering policy. This will require the producer to
+-- synchronize on the queue, send a doorbell when the queue has items in it.
+-- The queue can be an array which can be written directly to the IO device.
+--
+-- We can try both and see which one performs better.
+
+-- XXX add this example after fixing the serial stream rate control
+-- >>> S.toList $ S.take 5 $ intervalsOf 1 FL.sum $ constRate 2 $ S.enumerateFrom 1
+-- > [3,7,11,15,19]
+--
+-- | Group the input stream into windows of @n@ second each and then fold each
+-- group using the provided fold function.
+--
+-- @since 0.7.0
+{-# INLINE intervalsOf #-}
+intervalsOf
+    :: (IsStream t, MonadAsync m)
+    => Double -> Fold m a b -> t m a -> t m b
+intervalsOf n f m = FL.grouped (catMaybes f) s
+    where
+    s = S.map (\x -> (Just x, False)) m `parallel` S.repeatM timeout
+    timeout = do
+        liftIO $ threadDelay (round $ n * 1000000)
+        return (Nothing, True)
+    catMaybes = FL.lfilter isJust . FL.lmap fromJust